:recycle: refactor: 프로필 조회시 팔로잉 수, 찜 수 조회되도록 수정(#168)
